requirements.txt
streamlit>=1.20
langchain_groq
sentence-transformers>=2.2.2
chromadb>=0.3.32
pypdf>=3.8.0
python-dotenv
requests
beautifulsoup4